Use bash in place of mysql's "replace"
We used "replace" from the mysql package (which is actually an alias for
mariadb) to replace strings without having to worry about regexp
nonsense (e.g. if we used sed).

This was a pretty huge dependency for the sake of a single line; it was
fine if we have a NixOS machine using binary substitutes, but when
building everything from scratch it brought in a load of unnecessary
junk.