More foldls and speed up looking for top-level definitions
We need to process 'define-funs-rec' definitions, these may define multiple
(mutually-recursive) functions at once, and annoyingly they can be any
combination of mono/polymorphic. To make things easier, we just defer to
'names-in'.

However, this function turns out to be quite slow, so instead we bite the bullet
and actually check the names as we find them; after all, we've pattern-matched
far enough to see it's a 'define-funs-rec', so it's not much work to dig into
the names.