Try to use cached data more
We leave the generic functions in place, for testing and for the occasional
command that accepts a benchmark on stdin. Those times when we run expensive
commands on the TIP benchmarks (as specified in BENCHMARKS_FALLBACK), we do two
things:

 - Define a nullary procedure which reads the results from an environment
   variable, failing if it isn't set
 - Define a separate script which writes the desired result to stdout

We invoke these scripts using Nix (in mkCache) so the results are cached to disk
based on the input hash, and all that nice Nixy goodness. If one script requires
another env var to already be set, we supply it from the same cached
environment. When testing, we set all of the env vars, but we use a subset of
the benchmark files to ensure they remain reasonably fast.

This does have the unfortunate consequence that we can only invoke the tests
when the scripts are runnable, but that's not too bad:

 - We were previously not testing the "externally facing" scripts at all; we
   were pushing as much logic as possible into the libraries and only invoking a
   single procedure in each script, to keep the failure surface small.
 - If and when they fail, it will be with the small subset of inputs used by the
   tests, not a huge set of hundreds of files.
 - It should be possible, by setting env vars appropriately, to use no
   cache (e.g. /dev/null) and invoke only those tests which don't need
   them (e.g. using a regex)