Reduce usage of custom-ite, especially in custom-div
Isabelle is hanging when given definitions like mergesort, which uses custom-div
to find the mid-point of a list by dividing the length by 2.

After fiddling around, it looks like this custom-div call is causing the hang,
and changing custom-div to use pattern-matching instead of custom-ite fixes it.

The problem is probably due to something like strict evaluation: custom-ite is a
normal function, so strict evaluation will evaluate its arguments before
choosing which one to keep. Since one of these is a base-case and the other is a
recursive-case, we end up always recursing before ever getting the chance to
choose the base case.