Number variables from left to right, ignoring types
We were indexing per type, so (variable 0 "Bool") is distinct from
(variable 0 "Int"), but this is problematic if we don't know the types.

If we assume that we're only dealing with type-correct (or untyped)
expressions (garbage in, garbage out), and each expression has a principle type,
then we can ignore the types completely. We already do this for constants, but
they have unique, known names. For variables we now just number them distinctly,
so we don't have to know their type.

In particular, this solves the problem of having an equation like:

    func1 a b = func2 a

We can replace 'a' with '(variable 0 "unknown")', but what should we use for
variable 'b'? We could put '(variable 0 "unknown2")' or '(variable 1 "unknown")'
and which is correct depends on whether 'a' and 'b' have the same type. Even if
we're consistent, e.g. choosing "unknown" for both types, we hit problems if we
try to match the resulting parsed equation with one parsed from a typed
equation. If the typed version knows that 'a' and 'b' are distinct, we get a
mismatch; if we used the other scheme and gave each variable a fresh type, we'd
get a mismatch when they're the same (and, of course, all of the combinations
when there are more than 2 variables involved).

With the new scheme we will always just number from left to right. We could do
this during matching, but why not just do it everywhere?