Remove some of the indirection for tip-benchmark-smtlib
In particular, we take the standalone scripts/mk_final_defs.rkt script and put
its contents into the Nix build script instead.

To make this work, we add a little Nix function for turning the given Racket
code into a builder script (using 'wrap' to set the environment, the
'PLTCOLLECTS' env var to give access to the scripts/ dir and a shebang).